+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/AdaPots.hs
@@ -153,7 +153,8 @@ consumedTxBody ::
   Consumed
 consumedTxBody txBody pp dpstate utxo =
   Consumed
-    { conInputs = coinBalance (txInsFilter utxo (txBody ^. inputsTxBodyL))
+    { conInputs =
+        coinBalance (txInsFilter utxo (txBody ^. inputsTxBodyL))
     , conRefunds = certsTotalRefundsTxBody pp dpstate txBody
     , conWithdrawals = fold . unWithdrawals $ txBody ^. withdrawalsTxBodyL
     }